pyproject.toml
[tool.poetry]
name = "financial_tool"
version = "0.1.0"
description = "Starter scaffold for Financial Statement Generation Platform (FastAPI + SQLAlchemy)"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-multipart = "^0.0.9"
SQLAlchemy = "^2.0.25"
databases = "^0.9.0"       # FIXED: Updated to a valid existing version
pydantic = "^2.6"
pydantic-settings = "^2.1"
jinja2 = "^3.1"
weasyprint = "^63.0"
pandas = "^2.2"            # Stable version
openpyxl = "^3.1"
python-dotenv = "^1.0"
alembic = "^1.13"
psycopg2-binary = "^2.9"
aiosqlite = "^0.19.0"      # Added explicit dependency since you use sqlite
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "4.0.1"  # <--- FORCE THIS SPECIFIC VERSION

[tool.poetry.dev-dependencies]
pytest = "^8.0"
ruff = "^0.3.0"            # FIXED: Valid version
black = "^24.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"